requirements.txt
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
python-pptx==0.6.21
aixplain==0.2.33
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.10.6
